fixup! Add extractParams for Argon2, PBKDF2, and Scrypt
