+++ password/password.cabal
@@ -64,10 +64,10 @@ library
       Data.Password.Bcrypt
       Data.Password.PBKDF2
       Data.Password.Scrypt
-      Data.Password.Internal
       Data.Password.Validate
   other-modules:
       Paths_password
+      Data.Password.Internal
   build-depends:
       base        >= 4.9      && < 5
     , base64      >= 0.3      && < 0.5
@@ -116,7 +116,6 @@ test-suite password-tasty
     , PBKDF2
     , Scrypt
     , TestPolicy
-    , TestUtils
     , Validate
     , Paths_password
   ghc-options:
@@ -125,7 +124,6 @@ test-suite password-tasty
       base >=4.9 && <5
     , password
     , password-types
-    , base64
     , bytestring
     , cryptonite
     , memory
